MemeMe2.0
The MemeMe 2.0 is an app that enables a user to take a picture, and add text at the top and bottom to form a meme, and share that meme on Facebook and Twitter and also by SMS or email. In this project, sent memes appears in a tab view with two tabs: a table view and a collection view.